Separate umbrella header from class header
